requirements.txt
python-docx==1.1.2
pyinstaller==6.13.0
Pillow==10.4.0